add transaction model

+++ app/src/main/java/com/example/walletv1/model/HistoryItemModel.kt
@@ -3,7 +3,7 @@ package com.example.walletv1.model
 import com.google.gson.annotations.SerializedName
 
 data class HistoryItemModel(
-    @SerializedName("amount") val amount: Int,
+    @SerializedName("amount") val amount: Double,
     @SerializedName("category") val category: String,
     @SerializedName("confirmations") val confirmations: Int,
     @SerializedName("time") val time: Int,
